code/requirements.txt
requests==2.22.0
Pillow==6.2.0
boto3==1.9.248
botocore==1.12.248
s3transfer==0.2.1